requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
google-ai-generativelanguage==0.4.0
google-api-core==2.17.0
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
idna==3.6
numpy==1.26.4
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
proto-plus==1.23.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pytesseract==0.3.10
requests==2.31.0
rsa==4.9
soupsieve==2.5
tqdm==4.66.1
typing_extensions==4.9.0
urllib3==2.2.0
